refactor(frontend): rename SendTokenContext to TokenActionContext
The context component is used well beyond the Send flow (Withdraw, Stake/Unstake, Liquidium supply/repay, ETH WalletConnect, AI assistant), so the Send-specific name was misleading. Rename the component and update all 8 importers.

Pure rename — no behavior change. The underlying send.store context (SEND_CONTEXT_KEY / SendContext) is intentionally left untouched to keep the diff focused.
